+++ src/main/kotlin/team/b2/bingojango/domain/vote/service/VoteService.kt
@@ -11,7 +11,7 @@ import team.b2.bingojango.domain.vote.dto.request.VoteRequest
 import team.b2.bingojango.domain.vote.dto.response.VoteResponse
 import team.b2.bingojango.domain.vote.repository.VoteRepository
 import team.b2.bingojango.global.exception.cases.*
-import team.b2.bingojango.global.security.UserPrincipal
+import team.b2.bingojango.global.security.util.UserPrincipal
 import team.b2.bingojango.global.util.EntityFinder
 
 @Service
