+++ src/main/kotlin/team/b2/bingojango/global/security/jwt/TokenRepository.kt
@@ -1,8 +1,8 @@
-package team.b2.bingojango.global.security.jwt.repository
+package team.b2.bingojango.global.security.jwt
 
 import org.springframework.data.jpa.repository.JpaRepository
 import team.b2.bingojango.domain.user.model.User
-import team.b2.bingojango.global.security.jwt.model.RefreshToken
+import team.b2.bingojango.global.security.jwt.RefreshToken
 
 interface TokenRepository : JpaRepository<RefreshToken, Long> {
     fun findByUser(user: User): RefreshToken?
